requirements.txt
aiohttp==3.9.3
colorlog==6.8.2
gradio==3.41.2
httpx==0.27.0
librosa==0.10.1
numpy==1.26.4
psutil==5.9.8
PyYAML==6.0.1
Requests==2.31.0
rich==13.7.1
setuptools==69.2.0
torch==2.2.1+cu118
websockets==11.0.3
